refactored addressbook loading
